DJango/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
Pillow==10.1.0
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
rpds-py==0.13.1
setuptools==69.0.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
validate-docbr==1.10.0